backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.0
pyserial==3.5
python-socketio==5.12.1
pywebview==5.4
requests==2.32.3
werkzeug==3.1.3
eventlet==0.39.0
pyinstaller==6.11.1